Refactoring and starting with Project entity
