+++ src/main/java/ch/uzh/ifi/hase/soprafs24/models/user/User.java
@@ -1,4 +1,4 @@
-package ch.uzh.ifi.hase.soprafs24.models;
+package ch.uzh.ifi.hase.soprafs24.models.user;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
