+++ src/main/java/ch/uzh/ifi/hase/soprafs24/models/user/UserRegister.java
@@ -1,8 +1,9 @@
-package ch.uzh.ifi.hase.soprafs24.models;
+package ch.uzh.ifi.hase.soprafs24.models.user;
 
-public class UserUpdate {
+public class UserRegister {
     private String name;
     private String username;
+    private String email;
     private String password;
 
     public String getName() { return name; }
@@ -11,7 +12,9 @@ public class UserUpdate {
     public String getUsername() { return username; }
     public void setUsername(String username) { this.username = username; }
 
+    public String getEmail() { return email; }
+    public void setEmail(String email) { this.email = email; }
+
     public String getPassword() { return password; }
     public void setPassword(String password) { this.password = password; }
-
 }
