+++ src/main/java/ch/uzh/ifi/hase/soprafs24/repository/UserRepository.java
@@ -1,7 +1,9 @@
 package ch.uzh.ifi.hase.soprafs24.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import ch.uzh.ifi.hase.soprafs24.models.User;
+
+import ch.uzh.ifi.hase.soprafs24.models.user.User;
+
 import java.util.Optional;
 
 public interface UserRepository extends MongoRepository<User, String>
